+++ src/pygeoprocessing/__init__.py
@@ -19,6 +19,8 @@ from .geoprocessing import align_bbox
 from .geoprocessing import array_equals_nodata
 from .geoprocessing import build_overviews
 from .geoprocessing import calculate_disjoint_polygon_set
+from .geoprocessing import choose_dtype
+from .geoprocessing import choose_nodata
 from .geoprocessing import convolve_2d
 from .geoprocessing import create_raster_from_bounding_box
 from .geoprocessing import create_raster_from_vector_extents
@@ -33,6 +35,7 @@ from .geoprocessing import merge_bounding_box_list
 from .geoprocessing import new_raster_from_base
 from .geoprocessing import numpy_array_to_raster
 from .geoprocessing import raster_calculator
+from .geoprocessing import raster_map
 from .geoprocessing import raster_reduce
 from .geoprocessing import raster_to_numpy_array
 from .geoprocessing import rasterize