+++ src/pygeoprocessing/routing/watershed.pyx
@@ -628,10 +628,16 @@ def delineate_watersheds_d8(
         target_layer_name='watersheds'):
     """Delineate watersheds for a vector of geometries using D8 flow dir.
 
+    Note:
+        The ``d8_flow_dir_raster_path_band`` and ``outflow_vector_path`` files
+        must have the same spatial reference system.  The output watersheds
+        vector will use this same spatial reference system.
+
     Args:
         d8_flow_dir_raster_path_band (tuple): A (path, band_id) tuple
             to a D8 flow direction raster.  This raster must be a tiled raster
-            with block sizes being a power of 2.
+            with block sizes being a power of 2.  The output watersheds vector
+            will have its spatial reference copied from this raster.
         outflow_vector_path (str): The path to a vector on disk containing
             features with valid geometries from which watersheds will be
             delineated.  Only those parts of the geometry that overlap valid
@@ -701,7 +707,8 @@ def delineate_watersheds_d8(
 
     gtiff_driver = gdal.GetDriverByName('GTiff')
     flow_dir_srs = osr.SpatialReference()
-    flow_dir_srs.ImportFromWkt(flow_dir_info['projection_wkt'])
+    if flow_dir_info['projection_wkt']:
+        flow_dir_srs.ImportFromWkt(flow_dir_info['projection_wkt'])
 
     outflow_vector = gdal.OpenEx(outflow_vector_path, gdal.OF_VECTOR)
     if outflow_vector is None:
@@ -709,7 +716,8 @@ def delineate_watersheds_d8(
 
     driver = ogr.GetDriverByName('GPKG')
     watersheds_srs = osr.SpatialReference()
-    watersheds_srs.ImportFromWkt(flow_dir_info['projection_wkt'])
+    if flow_dir_info['projection_wkt']:
+        watersheds_srs.ImportFromWkt(flow_dir_info['projection_wkt'])
     watersheds_vector = driver.CreateDataSource(target_watersheds_vector_path)
     polygonized_watersheds_layer = watersheds_vector.CreateLayer(
         'polygonized_watersheds', watersheds_srs, ogr.wkbPolygon)
@@ -743,8 +751,6 @@ def delineate_watersheds_d8(
     LOGGER.info('Delineating watersheds')
     outflow_layer = outflow_vector.GetLayer()
     outflow_feature_count = outflow_layer.GetFeatureCount()
-    flow_dir_srs = osr.SpatialReference()
-    flow_dir_srs.ImportFromWkt(flow_dir_info['projection_wkt'])
     for feature in outflow_layer:
         # Some vectors start indexing their FIDs at 0.
         # The mask raster input to polygonization, however, only regards pixels
@@ -818,7 +824,8 @@ def delineate_watersheds_d8(
             gdal.GDT_UInt32,
             options=GTIFF_CREATION_OPTIONS)
         scratch_raster.SetGeoTransform(source_gt)
-        scratch_raster.SetProjection(flow_dir_info['projection_wkt'])
+        if flow_dir_info['projection_wkt']:
+            scratch_raster.SetProjection(flow_dir_info['projection_wkt'])
         # strictly speaking, there's no need to set the nodata value on the band.
         scratch_raster = None